refactor: move helper functions to utils and fix linting
